infra: extract `CompositeError` to use throughout the crate
