+++ packages/mdbook-trpl/src/lib.rs
@@ -33,3 +33,16 @@ pub fn parser(text: &str) -> Parser<'_> {
     opts.insert(Options::ENABLE_SMART_PUNCTUATION);
     Parser::new_ext(text, opts)
 }
+
+#[derive(Debug, thiserror::Error)]
+struct CompositeError(Vec<anyhow::Error>);
+
+impl std::fmt::Display for CompositeError {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        write!(
+            f,
+            "Error(s) rewriting input: {}",
+            self.0.iter().map(|e| format!("{e:?}")).collect::<String>()
+        )
+    }
+}
